12.0

+++ Framework-Headers/AccessibilityBundles/FaceTime.axbundle/Objc/DialerControllerAccessibility.h
@@ -14,9 +14,9 @@
 + (Class)safeCategoryBaseClass;
 + (id)safeCategoryTargetClassName;
 - (BOOL)_accessibilityHandwritingAttributeAcceptsRawInput;
-- (int)_accessibilityHandwritingAttributeAllowedCharacterSets;
+- (unsigned long long)_accessibilityHandwritingAttributeAllowedCharacterSets;
 - (BOOL)_accessibilityHandwritingAttributeCanDeleteCharacter;
-- (int)_accessibilityHandwritingAttributePreferredCharacterSet;
+- (unsigned long long)_accessibilityHandwritingAttributePreferredCharacterSet;
 - (void)_accessibilityInsertText:(id)arg1 atPosition:(long long)arg2;
 - (void)_accessibilityLoadAccessibilityInformation;
 - (void)_accessibilityReplaceCharactersAtCursor:(unsigned long long)arg1 withString:(id)arg2;
