+++ Framework-Headers/AccessibilityBundles/ManagedConfigurationUI.axbundle/Objc/MCQuestionViewAccessibility.h
@@ -14,9 +14,9 @@
 + (Class)safeCategoryBaseClass;
 + (id)safeCategoryTargetClassName;
 - (BOOL)_accessibilityHandwritingAttributeAcceptsRawInput;
-- (int)_accessibilityHandwritingAttributeAllowedCharacterSets;
+- (unsigned long long)_accessibilityHandwritingAttributeAllowedCharacterSets;
 - (BOOL)_accessibilityHandwritingAttributeCanDeleteCharacter;
-- (int)_accessibilityHandwritingAttributePreferredCharacterSet;
+- (unsigned long long)_accessibilityHandwritingAttributePreferredCharacterSet;
 - (BOOL)_accessibilityHandwritingAttributeShouldEchoCharacter;
 - (BOOL)_accessibilityHandwritingAttributeShouldPlayKeyboardSecureClickSound;
 - (void)_accessibilityInsertText:(id)arg1 atPosition:(long long)arg2;
