+++ Framework-Headers/PrivateFrameworks/iWorkImport._framework/Objc/TSCETableTransposedInfo.h
@@ -4,16 +4,16 @@
 //  Copyright (C) 1997-2019 Steve Nygard.
 //
 
-#import <Foundation/NSObject.h>
+#import <objc/NSObject.h>
 
 __attribute__((visibility("hidden")))
 @interface TSCETableTransposedInfo : NSObject
 {
     struct TSCERangeCoordinate _transposedBodyRange;
-    unsigned short _numberOfFooterRows;
+    unsigned int _numberOfFooterRows;
 }
 
-@property (nonatomic) unsigned short numberOfFooterRows; // @synthesize numberOfFooterRows=_numberOfFooterRows;
+@property (nonatomic) unsigned int numberOfFooterRows; // @synthesize numberOfFooterRows=_numberOfFooterRows;
 @property (nonatomic) struct TSCERangeCoordinate transposedBodyRange; // @synthesize transposedBodyRange=_transposedBodyRange;
